feat: Implemented role scanning logic

+++ src/ps2/service/ps2.game.scanning.service.ts
@@ -17,10 +17,10 @@ interface ChangesInterface {
 @Injectable()
 export class PS2GameScanningService implements OnApplicationBootstrap {
   private readonly logger = new Logger(PS2GameScanningService.name);
-  private messagesMap: Map<string, Message> = new Map();
   private charactersMap: Map<string, CensusCharacterWithOutfitInterface> = new Map();
   private changesMap: Map<string, ChangesInterface> = new Map();
-  private suggestionsMap: Map<string, ChangesInterface> = new Map();
+  private suggestionsMap: Map<string, ChangesInterface[]> = new Map();
+  private suggestionsCount = 0;
 
   constructor(
     private readonly censusService: CensusApiService,
@@ -34,28 +34,21 @@ export class PS2GameScanningService implements OnApplicationBootstrap {
 
   reset() {
     this.logger.log('Resetting maps...');
-    this.messagesMap.clear();
     this.charactersMap.clear();
     this.changesMap.clear();
     this.suggestionsMap.clear();
+    this.suggestionsCount = 0;
   }
 
-  async startScan(interaction: ChatInputCommandInteraction) {
+  async startScan(interaction: ChatInputCommandInteraction, dryRun = false) {
     const message = await interaction.channel.send('Starting scan...');
 
-    this.messagesMap.set(interaction.id, message);
-
-    await this.scanForMembership(message);
-  }
-
-  async scanForMembership(message: Message) {
     // Pull the list of verified members from the database and check if they're still in the outfit
     // If they're not, remove the verified role from them and any other PS2 Roles
     // Also send a message to the #ps2-leadership channel to denote this has happened
 
     const outfitMembers = await this.ps2MembersRepository.findAll();
     const length = outfitMembers.length;
-
     const characterPromises: Promise<CensusCharacterWithOutfitInterface>[] = [];
 
     for (const member of outfitMembers) {
@@ -68,10 +61,10 @@ export class PS2GameScanningService implements OnApplicationBootstrap {
 
     try {
       await message.edit(`Checking ${length} characters for membership status...`);
-      await this.removeLeavers(characters, outfitMembers, message);
+      await this.removeLeavers(characters, outfitMembers, message, dryRun);
 
-      // await message.edit(`Checking ${length} characters for role inconsistencies...`);
-      // await this.checkForSuggestions(characters, outfitMembers, message);
+      await message.edit(`Checking ${length} characters for role inconsistencies...`);
+      await this.checkForSuggestions(characters, outfitMembers, message);
     }
     catch (err) {
       console.log(err);
@@ -81,17 +74,17 @@ export class PS2GameScanningService implements OnApplicationBootstrap {
     }
 
     if (this.changesMap.size === 0) {
-      await message.edit('No changes made.');
+      await message.channel.send('✅ No automatic changes were performed.');
       this.logger.log('No changes were made.');
-      return this.reset();
     }
     else {
-      await message.edit(`There are currently ${outfitMembers.length} members on record. \n## 📝 ${this.changesMap.size} changes made`);
+      await message.channel.send(`## 📝 ${this.changesMap.size} change(s) made`);
       this.logger.log(`Sending ${this.changesMap.size} changes to channel...`);
     }
 
     for (const change of this.changesMap.values()) {
-      await message.channel.send(change.change);
+      const fakeMessage = await message.channel.send('dummy'); // Send a fake message first so it doesn't ping people
+      await fakeMessage.edit(change.change);
     }
 
     if (this.suggestionsMap.size === 0) {
@@ -99,19 +92,23 @@ export class PS2GameScanningService implements OnApplicationBootstrap {
       this.logger.log('No suggestions were made.');
     }
     else {
-      await message.channel.send('## 👀 Suggestions of changes to make');
-      this.logger.log(`Sending ${this.suggestionsMap.size} suggestions to channel...`);
-
+      await message.channel.send(`## 👀 ${this.suggestionsCount} manual correction(s) to make`);
+      this.logger.log(`Sending ${this.suggestionsCount} suggestions to channel...`);
     }
 
     for (const change of this.suggestionsMap.values()) {
-      await message.channel.send(change.change);
+      for (const suggestion of change) {
+        const fakeMessage = await message.channel.send('dummy'); // Send a fake message first so it doesn't ping people
+        await fakeMessage.edit(suggestion.change);
+      }
     }
+
+    await message.edit(`ℹ️ There are currently ${outfitMembers.length} members on record.`);
 
     return this.reset();
   }
 
-  async removeLeavers(characters: CensusCharacterWithOutfitInterface[], outfitMembers: PS2MembersEntity[], message: Message) {
+  async removeLeavers(characters: CensusCharacterWithOutfitInterface[], outfitMembers: PS2MembersEntity[], message: Message, dryRun = false) {
     // Save all the characters to a map we can easily pick out later
     for (const character of characters) {
       this.charactersMap.set(character.character_id, character);
@@ -127,13 +124,17 @@ export class PS2GameScanningService implements OnApplicationBootstrap {
         discordMember = await message.guild.members.fetch({ user: member.discordId, force: true });
       }
       catch (err) {
-        // No discord memember means they've left the server
+        // No discord member means they've left the server
         this.logger.log(`User ${character.name.first} has left the server`);
-        await this.ps2MembersRepository.removeAndFlush(member);
+
+        if (!dryRun) {
+          await this.ps2MembersRepository.removeAndFlush(member);
+        }
+
         this.changesMap.set(member.characterId, {
           character,
           discordMember: null,
-          change: `🫥️ Discord member for Character **${character.name.first}** has left the DIG server. Their verification status has been removed.`,
+          change: `- 🫥️ Discord member for Character **${character.name.first}** has left the DIG server. Their verification status has been removed.`,
         });
         continue;
       }
@@ -155,31 +156,36 @@ export class PS2GameScanningService implements OnApplicationBootstrap {
         const hasRole = discordMember.roles.cache.has(rankMap.discordRoleId);
 
         if (!hasRole) {
-          console.log(`User ${discordMember.user.username} (${discordMember.user.id}) does not have role "${role.name} to remove.`);
+          console.log(`User <@${discordMember.id}> does not have role "${role.name} to remove.`);
           continue;
         }
 
-        try {
-          await discordMember.roles.remove(rankMap.discordRoleId);
-        }
-        catch (err) {
-          await message.channel.send(`ERROR: Unable to remove role "${role.name}" from ${character.name.first} (${character.character_id}). Pinging <@${this.config.get('app.discord.ownerId')}>!`);
+        if (!dryRun) {
+          try {
+            await discordMember.roles.remove(rankMap.discordRoleId);
+          }
+          catch (err) {
+            await message.channel.send(`ERROR: Unable to remove role "${role.name}" from ${character.name.first} (${character.character_id}). Pinging <@${this.config.get('app.discord.ownerId')}>!`);
+          }
         }
       }
 
-      await this.ps2MembersRepository.removeAndFlush(member);
+      if (!dryRun) {
+        await this.ps2MembersRepository.removeAndFlush(member);
+      }
 
       this.changesMap.set(member.characterId, {
         character,
         discordMember,
-        change: `👋 <@${discordMember.id}>'s character **${character.name.first}** has left the outfit. Their roles and verification status have been stripped.`,
+        change: `- 👋 <@${discordMember.id}>'s character **${character.name.first}** has left the outfit. Their roles and verification status have been stripped.`,
       });
     }
   }
 
   async checkForSuggestions(characters: CensusCharacterWithOutfitInterface[], outfitMembers: PS2MembersEntity[], message: Message) {
-    console.log(this.changesMap);
     // Check if there are any characters in the outfit that have invalid discord permissions
+
+    const rankMap: RankMapInterface = this.config.get('app.ps2.rankMap');
 
     for (const member of outfitMembers) {
       // If already in the change set, they have been removed so don't bother checking
@@ -188,33 +194,62 @@ export class PS2GameScanningService implements OnApplicationBootstrap {
       }
 
       const character = this.charactersMap.get(member.characterId);
-      const rankMap: RankMapInterface = this.config.get('app.ps2.rankMap');
-      const currentRoles = await message.guild.members.fetch({ user: member.discordId, force: true });
+      const discordMember = await message.guild.members.fetch({ user: member.discordId, force: true });
 
-      // Get a filtered list of roles that are in the rank map
-      const filteredRoles = currentRoles.roles.cache.filter(role => Object.values(rankMap).some(rank => rank.discordRoleId === role.id));
+      console.log(`Checking roles of character ${character.name.first} (${character.character_id})`);
 
-      // If they don't have any roles according to the rank map, we can't check them anyway as we have no idea what to look for
-      if (!filteredRoles.size) {
-        continue;
-      }
+      // First get their rank
+      const rank = character.outfit_info?.rank_ordinal;
 
-      // Loop the now filtered roles and check if they have the correct role for their rank
-      for (const role of filteredRoles.values()) {
-        const rankRole = Object.values(rankMap).find(rank => rank.discordRoleId === role.id);
+      console.log('rank', rank);
 
-        // If the role is correct, continue
-        if (rankRole.rank === character.outfit_info.rank_ordinal) {
-          continue;
+      // Line their rank up with the correct role(s)
+      const shouldHaveRoles = Object.values(rankMap).filter((role) => role.rank === rank);
+
+      shouldHaveRoles.forEach((role) => {
+        // Get the role from the guild
+        const guildRole = message.guild.roles.cache.get(role.discordRoleId);
+
+        // Now check if the user has the role
+        const hasRole = discordMember.roles.cache.has(role.discordRoleId);
+
+        if (!hasRole) {
+          // Get all current suggestions, if any
+          const suggestions = this.suggestionsMap.get(member.characterId) || [];
+          suggestions.push({
+            character,
+            discordMember,
+            change: `- 😳 <@${discordMember.id}> is missing the role \`${guildRole.name}\``,
+          });
+          this.suggestionsMap.set(member.characterId, suggestions);
+          this.suggestionsCount++;
         }
+      });
 
-        // If the role is incorrect, add it to the suggestions map
-        this.suggestionsMap.set(member.characterId, {
-          character,
-          discordMember: currentRoles,
-          change: `❓ ${character.name.first} (${character.character_id}) has incorrect role "${role.name}"`,
-        });
-      }
+      // Now check if they have any roles they shouldn't have
+      const shouldNotHaveRoles = Object.values(rankMap).filter((role) => role.rank !== rank);
+
+      shouldNotHaveRoles.forEach((role) => {
+        // Get the role from the guild
+        const guildRole = message.guild.roles.cache.get(role.discordRoleId);
+
+        // Now check if the user has the role
+        const hasRole = discordMember.roles.cache.has(role.discordRoleId);
+
+        if (hasRole) {
+          // Get all current suggestions, if any
+          const suggestions = this.suggestionsMap.get(member.characterId) || [];
+          suggestions.push({
+            character,
+            discordMember,
+            change: `- 🤔 <@${discordMember.id}> has the role \`${guildRole.name}\` when their rank suggests they shouldn't!`,
+          });
+          this.suggestionsMap.set(member.characterId, suggestions);
+          this.suggestionsCount++;
+        }
+      });
+
+      console.log('shouldHaveRoles', shouldHaveRoles);
     }
   }
 }
